+++ src/main/java/io/github/queerbric/inspecio/InspecioConfig.java
@@ -29,10 +29,7 @@ import net.minecraft.util.math.MathHelper;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
+import java.nio.file.*;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
@@ -703,7 +700,7 @@ public class InspecioConfig {
 	private static InspecioConfig backupAndRestore(InspecioConfig config) {
 		try {
 			if (createConfigBackupDirectoryIfNeeded())
-				Files.copy(CONFIG_PATH, CONFIG_BACKUP_PATH);
+				Files.copy(CONFIG_PATH, CONFIG_BACKUP_PATH, StandardCopyOption.REPLACE_EXISTING);
 
 			config.save();
 		} catch (IOException e) {